refactor: Move address storage from DO to shared KV utility
